Fresh whole-branch review findings: wrap LMoveCount's ordering, move decoders to companion objects, fix Long/Int count truncation, LCS byte-length, ROLE sentinel case
A fresh, from-scratch code review of the whole 9732cdc4..tip diff (no
prior context, unlike the agents that wrote each piece) surfaced real
issues nothing individual-PR review had caught:

- LMoveCount.UpTo/Exactly embedded raw Lettuce LMovemArgs.Ordering
  directly, violating the "always wrap Lettuce types" rule adopted
  mid-effort - forced TestScenarios.scala to import Lettuce directly.
  New native LMoveOrdering ADT.
- Decode logic for RedisRole/SlowLogEntry/CommandInfo/RedisServerTime/
  TrackingInfo still lived as private methods on BaseRedis, even
  though the most recent PR in this series established "decode logic
  lives in the target type's own companion object" for XStreamInfo/
  XGroupInfo/XConsumerInfo. All five moved to match.
- zmPopMin/zmPopMax silently truncated their public Long count via
  .toInt before calling Lettuce's zmpop, which only exposes an
  Int-count overload - unlike bzmPopMin/bzmPopMax (Long count, correct,
  since bzmpop's Lettuce overload genuinely takes Long). Changed the
  public signature to Int instead of silently truncating a Long a
  caller might reasonably pass. Same fix for lPos's count-taking
  overloads (Lettuce's lpos is Int-only, no Long overload exists).
- toLcsResult's plain (non-IDX) length was computed via Scala's
  String#length() (UTF-16 code units) instead of the byte count Redis
  actually reports (and lcsLen returns) - silently wrong for any
  non-ASCII matched content. Fixed to re-encode via UTF-8 and count
  bytes.
- toRedisRole had no case for a Sentinel instance's ROLE reply
  ("sentinel", [master names]) - a real, documented reply shape this
  trait's own doc comment claimed to model completely. New
  RedisRole.Sentinel case.
- ServerCommands[F, K] -> [F, K, V] (from #1193) turned out to be
  unnecessary: no method in the whole ServerCommands trait family
  actually needs V (RedisServerTime has no V field; TIME's V-typed
  Lettuce call only matters inside BaseRedis's concrete implementation,
  which already has its own V in scope as a class type parameter).
  Reverted the arity widening - pure unused surface area.
- Added a comment explaining KillArgs.Builder.typeSlave() vs
  ClientListArgs.Builder.typeReplica() for the same ClientType.Replica
  case: not a copy-paste slip, KillArgs.Builder's Lettuce API genuinely
  lacks a typeReplica() alias.
- Test coverage: objectRefcount's actual (verified empirically)
  missing-key behavior - returns 0, not an error, since Lettuce's own
  decoder collapses this before it ever reaches redis4cats.

Verified: RedisSpec + RedisClusterSpec both green (36/36) against a
real 8.10.1 cluster + single-node/replica, sbt mdoc clean.

+++ modules/effects/src/main/scala/dev/profunktor/redis4cats/algebra/lists.scala
@@ -63,8 +63,10 @@ trait ListGetter[F[_], K, V] {
   def lRange(key: K, start: Long, stop: Long): F[List[V]]
   def lPos(key: K, value: V): F[Option[Long]]
   def lPos(key: K, value: V, args: LPosArgs): F[Option[Long]]
-  def lPos(key: K, value: V, count: Long): F[List[Long]]
-  def lPos(key: K, value: V, count: Long, args: LPosArgs): F[List[Long]]
+
+  /** `count` is `Int`, not `Long` - Lettuce's multi-match `LPOS` only exposes an `Int`-count overload. */
+  def lPos(key: K, value: V, count: Int): F[List[Long]]
+  def lPos(key: K, value: V, count: Int, args: LPosArgs): F[List[Long]]
 }
 
 trait ListSetter[F[_], K, V] {
